+++ src/main/java/nextstep/domain/player/Player.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.player;
 
 import java.util.Objects;
