+++ src/test/java/nextstep/domain/player/NameTest.java
@@ -1,5 +1,6 @@
-package nextstep.domain;
+package nextstep.domain.player;
 
+import nextstep.domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
